made sophomorix-group --kill work again

+++ sophomorix-samba/modules/SophomorixSambaAD.pm
@@ -724,34 +724,49 @@ sub AD_group_kill {
 
     &Sophomorix::SophomorixBase::print_title("Killing group $group ($type, $school):");
     my ($count,$dn_exist,$cn_exist)=&AD_object_search($ldap,$root_dse,"group",$group);
-    if ($count > 0){
 
-        # deleting share
-        if ($smb_share ne  "unknown"){
-            my $smb = new Filesys::SmbClient(username  => $DevelConf::sophomorix_administrator,
-                                             password  => $smb_admin_pass,
-                                             debug     => 1);
-            # trying to delete homes (success only if it is empty)
-            my $smb_share_homes=$smb_share."/homes";
-            my $return1=$smb->rmdir($smb_share_homes);
-            if($return1==1){
-                print "OK: Deleted empty dir with succes $smb_share_homes\n"; # smb://linuxmuster.local/<school>/subdir1/subdir2
-                # go on an recursively delete group/share and
-                my $return2=$smb->rmdir_recurse($smb_share);
-                if($return2==1){
-                    print "OK: Deleted with succes $smb_share\n"; # smb://linuxmuster.local/<school>/subdir1/subdir2
-                    # deleting the AD account
-                    my $command="samba-tool group delete ". $group;
-                    print "   # $command\n";
-                    system($command);
+    print "Here $type\n";
+    if ($count > 0){
+        if ($type eq "adminclass"){
+            ### adminclass #####################################
+            # deleting share if possible, when succesful  the account
+            if ($smb_share ne  "unknown"){
+                my $smb = new Filesys::SmbClient(username  => $DevelConf::sophomorix_administrator,
+                                                 password  => $smb_admin_pass,
+                                                 debug     => 1);
+                # trying to delete homes (success only if it is empty)
+                my $smb_share_homes=$smb_share."/homes";
+                my $return1=$smb->rmdir($smb_share_homes);
+                if($return1==1){
+                    print "OK: Deleted empty dir with succes $smb_share_homes\n"; # smb://linuxmuster.local/<school>/subdir1/subdir2
+                    # go on an recursively delete group/share and
+                    my $return2=$smb->rmdir_recurse($smb_share);
+                    if($return2==1){
+                        print "OK: Deleted with succes $smb_share\n"; # smb://linuxmuster.local/<school>/subdir1/subdir2
+                        # deleting the AD account
+                        my $command="samba-tool group delete ". $group;
+                        print "   # $command\n";
+                        system($command);
+                    } else {
+                        print "ERROR: rmdir_recurse $smb_share $!\n"; # smb://linuxmuster.local/<school>/subdir1/subdir2
+                    }
                 } else {
-                    print "ERROR: rmdir_recurse $smb_share $!\n"; # smb://linuxmuster.local/<school>/subdir1/subdir2
+                    print "ERROR: rmdir $smb_share_homes $!\n"; # smb://linuxmuster.local/<school>/subdir1/subdir2
                 }
-            } else {
-                print "ERROR: rmdir $smb_share_homes $!\n"; # smb://linuxmuster.local/<school>/subdir1/subdir2
+
             }
-       }
-       return;
+
+	} elsif ($type eq "project"){
+
+
+	} elsif ($type eq "sophomorix-group"){
+                ### sophomorix-group #####################################
+                # there is no share, just delete the group
+                my $command="samba-tool group delete ". $group;
+                print "   # $command\n";
+                system($command);
+        }
+        return;
     } else {
        print "   * Group $group nonexisting ($count results)\n";
        return;
